+++ test/pieces.js
@@ -351,8 +351,6 @@ describe('Pieces', function() {
     const user = apos.user.newInstance();
     assert(user);
 
-    user.firstName = 'ad';
-    user.lastName = 'min';
     user.title = 'admin';
     user.username = 'admin';
     user.password = 'admin';
@@ -362,8 +360,6 @@ describe('Pieces', function() {
     await apos.user.insert(apos.task.getReq(), user);
 
     const user2 = apos.user.newInstance();
-    user2.firstName = 'ad';
-    user2.lastName = 'min2';
     user2.title = 'admin2';
     user2.username = 'admin2';
     user2.password = 'admin2';