+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## UNRELEASED
 
+### Breaks
+
+* Removes the `firstName` and `lastName` fields in user pieces.
+
 ### Adds
 
 * Major performance improvement to the `nestedModuleSubdirs` option.
@@ -15,6 +19,7 @@
 * Contributors can submit drafts for review ("Propose Changes" versus "Publish Changes").
 * Editors and admins can manage submitted drafts.
 * Support for virtual piece types, such as submitted drafts, which in actuality manage more than one type of doc.
+* Sets `username` fields to follow the user `title` field to remove an extra step in user creation.
 
 ### Fixes